Adding archive handler and compression handler classes.
Splitting responsibility for handling archives and compression out from the
recursor closure and into its own separate classes respectively.

I know we were originally going to stick to closures and hold out on going
full OOP on the core until v2, and 'cleanup' won't work on classes, but
seeing as I need to completely overhaul archive functionality already
anyway, and I don't feel too great about doing it all again 6 months or
whatever down to track if I can avoid it, I've opted to take this approach
immediately for archives and compression, and we can just consider it a
'stepping stone' to a future v2 or whatever.

Feel free to review/audit/comment this commit if you'd like. I don't shy
away from feedback! However, don't pull this into master just yet. This
commit forks a new branch, and a crapload of work still needs to be done
yet to the phpMussel core before anything is viable. Committing the work
done thus far now to avoid monster-sized commits later down the track, when
we're ready to pull (this commit is only a small piece of what's needed).