Remove crowdin configuration file.
A really cool service, and definitely could be useful, if I was to put
enough time into it and manage it correctly, but I've had the integration
set up for a few months now and yet still haven't gotten around to setting
it up properly or managing it properly, and yet have been continuing to
push L10N updates manually anyway, so I doubt, at this point, that I'll end
up getting around to it, and it'll likely be much easier in the long-term
to just continue doing it manually and not worry about integration with any
platforms for this purpose. Removing the configuration file accordingly.